[API-630] Fix Multiple/Array Joining
Before, it was selecting the individual fields in to the result, which was then being filtered. This worked for certain example use cases like the employee_habit, where we get back a string array of habits. But it worked very obtusely. Now when you have a type: Array field joined on, it'll return the array of model instances. Much betterer, I think.

https://jira.appcelerator.org/browse/API-630